This is a proof of concept pull request to gather initial feedback. It should not merge and close after some discussion. I’m mostly looking for comments on whether this is viable right now.

The change applies a simple heuristic for excluding some source files from dependencies when tsserver traverses a program’s module dependency graph. If a dependency “re-exports” a file with the `export { ... } from "..."` syntax, the file of that export will not be added to `Program` if none of the identifiers are used.

Suppose a package `example `uses only `SOME_CONST_A` from `example-dep`.

```ts
// packages/example-dep/index.ts
export { SomeComponentA } from "./components/SomeComponentA";
export { SomeComponentB } from "./components/SomeComponentB";
export { someUtilA } from "./utils/someUtilA";
export { someUtilB } from "./utils/someUtilB";
export { SOME_CONST_A, SOME_CONST_B } from "./components/consts";
```

```ts
// packages/example/index.ts
import { SOME_CONST_A } from "@ref/example-dep"
```

```ts
// packages/example/tsconfig.json
{
  "compilerOptions": {
    "composite": true,
    "baseUrl": "./",
    "paths": {
       "@ref/a": [ "../a" ]
    }
  },
  "references": [
    "path": "../example-dep"
  ]
}
```

In this case files from `components` and `utils` are no longer added. This can create big savings in memory and performance if `components`/`utils` loaded large dependencies themselves.

## Current Issues

- Elided identifiers no longer appear in completion responses. In the example above, SomeComponentA would not show while working on packages/example/index.ts. (More on this later.)
- The current implementation in this PR doesn’t add elided files to the program later if the identifier is added.
- Haven’t written any new tests for this yet.

## Questions

I’m curious if the TypeScript team is willing to work with me on this change. Most of the problems with this involve handling identifiers defined in elided source files.

1. After this change, identifiers from elided source files don’t appear in completion responses. Do they need to? (I assume the answer is yes, but wanted to verify.)
    1. Despite not appearing in completions, they do appear in the "Quick Fix..." menu for auto-imports.
2. If the elided identifiers are re-added, they won’t have full typing information in the completion prompt initially. We could possibly fetch this when a user hovers hover the entry though. Is that an acceptable trade-off?
3. Is the function call stack the best way to carry over `Identifier`s being searched for? This made calls to `processImportedModules` a bit ugly.