@amcasey 2020-10-08
Reminds me of this change: https://github.com/aws/aws-sdk-js/pull/3398