@amcasey 2020-10-08
I agree that being able to pull in some files without pulling in all files is a huge perf win, but I'd like to hear from someone more knowledgeable that this is a good way to achieve it.  I would have guessed it had more to do with the shape of the library being imported.

@gluxon Do you have numbers?  Did you apply your change to a sample project and see a big perf win?