@gluxon 2020-10-08
Thanks for giving initial thoughts @amcasey!

> I agree that being able to pull in some files without pulling in all files is a huge perf win, but I'd like to hear from someone more knowledgeable that this is a good way to achieve it.

Happy to wait for additional thoughts. The way I did it in this PR definitely needs improvements. I wrote the V0 here quickly to see what the perf differences would be and to check whether obvious problems would get caught by tests.

There don't seem to be any tests failures, which I was surprised by. I think this is mostly because the files elided here are coming from project references which didn't get [source code added until TypeScript 3.7 recently](https://devblogs.microsoft.com/typescript/announcing-typescript-3-7/#build-free-editing-with-project-references).

>  I would have guessed it had more to do with the shape of the library being imported.

Definitely. I should add that this improved editor startup times in our monorepo specifically because we have a lot of `index.ts` files from packages re-exporting its internals. Commonly new packages import from a shared react components package but would only need a few of the many in there.

> @gluxon Do you have numbers? Did you apply your change to a sample project and see a big perf win?

There's a large package on my team that got its `createProgram` time cut to 1/3 of the original time. (~30s to ~11s). Another large package we've had problems with went from ~28s to ~20s. I didn't give numbers initially because the differences depend heavily on project reference structure.