@andrewbranch 2020-10-08
It’s pretty cool that you were able to get a proof of concept working with such a small change, and it would be awesome if we could safely do something like this. Unfortunately, it could very easily break a build, and there’s (currently) no way to know whether or not it _would_ in a particular instance without doing most of the work of a full build.

There is an exact analogy to how bundlers opt into this same behavior of dead code elimination or live code inclusion—because loading JavaScript modules may have important side effects, they can’t do this in general; they have to know that the code they are omitting is “pure.” So, they read the `pure` field in package.json to opt into this behavior on a per-package basis. Similarly, loading a TypeScript file into a program may have side effects on the program state: the file may introduce globals, ambient module declarations, or module augmentations. In your example, it’s clear that only `SOME_CONST_A` is _imported_, but that isn‘t sufficient to tell us that it’s the only thing from `@ref/example-dep` that is _used_. If we omitted all the other modules imported by the index file, it’s very possible that we would be omitting global declarations that are referenced in files we _are_ including.

Of course, you’re absolutely right that this behavior would _often_ work and would often have significant benefits. It makes me wonder if, extending the bundler analogy, we could let individual packages indicate that their typings are “pure,” and enable this behavior on a per-package basis. If we had something like that, we’d probably want the authors of those packages to be able to (or _have_ to) validate their claim that their declarations are side-effect-free: modules would not be allowed to import scripts, declare globals, or augment other modules.

It’s also worth clarifying that what you’ve done here so far is _not_ specific to TS Server, despite the PR title—which is a good thing. Including a different set of modules between `tsc` and the language service / TS Server is a non-starter. (You’d wind up in situations where your editor shows errors but command-line compilation fails, or vice versa.) As another implementation note (which is probably not super relevant at this stage), suppose that your index file used a series of `export * from`, which I think is more common for barrels, instead of using export specifiers—I can’t tell at a glance if this proof of concept would handle that, but I think perhaps it wouldn’t? I think in a hypothetical world where we could implement something like this, the implementation _would_ get a bit more complicated. It’s an interesting experiment though, so thanks for getting a conversation started!