@gluxon 2020-10-13
Thanks for that fantastic in-depth explanation. We've been relying on webpack tree-shaking to reduce bundle sizes on my team, so this PR was definitely inspired by that. I wasn't aware of side-effects to the typing system through globals and module augmentations, but I see it's well-documented on typescriptlang.org and makes sense. Thanks for calling that out. It did feel like I was missing something that made this PR too easy. 😅

> Of course, you’re absolutely right that this behavior would often work and would often have significant benefits. It makes me wonder if, extending the bundler analogy, we could let individual packages indicate that their typings are “pure,” and enable this behavior on a per-package basis. If we had something like that, we’d probably want the authors of those packages to be able to (or have to) validate their claim that their declarations are side-effect-free: modules would not be allowed to import scripts, declare globals, or augment other modules.

I'd love to explore this if the team is open to it. It sounds like the specific requirements here are to add a new flag and error on side-effectful TypeScript statements if they're seen. I'll file a new issue linking your comment to discuss that prerequisite further.

> It’s also worth clarifying that what you’ve done here so far is not specific to TS Server, despite the PR title—which is a good thing. Including a different set of modules between tsc and the language service / TS Server is a non-starter. (You’d wind up in situations where your editor shows errors but command-line compilation fails, or vice versa.)

Yes, you're right. I was specifically aiming for a tsserver improvement but this does affect `tsc` after rethinking. My thought was that `tsc` would always load root names, but dependencies are still affected in general.

Ensuring errors are consistent between compilation and editing makes sense. 👍

> As another implementation note (which is probably not super relevant at this stage), suppose that your index file used a series of export * from, which I think is more common for barrels, instead of using export specifiers—I can’t tell at a glance if this proof of concept would handle that, but I think perhaps it wouldn’t? I think in a hypothetical world where we could implement something like this, the implementation would get a bit more complicated.

Yup, you're right again. This PR doesn't handle `export * from` syntax because that would require peeking into the file itself (recursively) to see what it exports. The PR would be much larger if it handled that.

>  It’s an interesting experiment though, so thanks for getting a conversation started!

I really appreciate you and everyone else being receptive. This is a great open source community. Thanks again.