@gluxon 2020-10-14
> I'll file a new issue linking your comment to discuss that prerequisite further.

https://github.com/microsoft/TypeScript/issues/41104

Going to close this experiment for now.